requirements.txt
gcsa==2.3.0
dateparser==1.2.0
pytz==2024.1
groq==0.9.0
gradio==4.31.3
python-dotenv==1.0.1